Completely overhauling the rendering system. Implementing proper exception handing, strict typing to shader terms, more flexible and encapsulated binding of resources and more.
After done refactoring the system, will start to implement the spatial partitioning and get the GUI items back up and running. After GUI, will implement sprites to be used as place holders, if proper key-frame animated 3D models become feasible.

Spatial partitioning of the static world mesh is crucial however, if a proper hit detection is to be implemented.