Implementing the alternative scene loading infrastructure now. A new class Worldmesh will be created soon, and it will essentially house the entire static worldgeometry in itself. It shall contain representation for both rendering and intersection queries.
Calling the resulting piles of triangles as "mashes" instead of "meshes".

Not sure whether duplicating vertex data in Oct-tree will be a big deal performance wise. For N-level partitioning, the spatial requirement goes O(N).

Each node of the tree should either refer to, or contain explicitly all the triangles belonging to its area. In essence, all the triangles of its sub nodes.

The triangles associated with each node should also be sorted by material attributes as to minimize the state changes by the OpenGL device.